pyproject.toml
[tool.poetry]
name = "telegram-scrapper"
version = "0.1.0"
description = ""
authors = ["d0gied <german.n.u@mail.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pyrogram = "^2.0.106"
aiogram = "^3.8.0"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.1"
python-levenshtein = "^0.25.1"

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
